requirements.txt
### 7. `requirements.txt`

imapclient
email-validator
pandas
sqlalchemy
openai
python-dotenv